requirements.txt
aiogram==3.13.1
aiohttp==3.10.5
asyncpg==0.29.0
gspread==6.1.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.151.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1


